requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
idna==3.7
numpy==1.26.4
opencv-python-headless==4.10.0.84
packaging==24.1
pillow==10.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyyaml==6.0.1
requests==2.32.3
six==1.16.0
typing-extensions==4.12.2
urllib3==2.2.2
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
gunicorn==21.2.0
opencv-contrib-python-headless==4.10.0.84
